MAJOR CHANGE TO WINDOWS LINKING - READ!!
I have now configured the windows version to link against the libiomp5md.lib/dll that is shipped with Intel OneApi. We were previously linking against the version shipped with LLVM, but it was removed in latest release. This requires you to change your path and environment variables as follows.

1. Assuming the default oneAPi install location. Add the following to your path

   C:\Program Files (x86)\Intel\oneAPI\compiler\latest\windows\redist\intel64_win\compiler

2. Set the environment variable ONEAPI_ROOT. Assuming a standard install it should be as follows. It may have already been
set during your installation.

    ONAPI_ROOT = C:\Program Files (x86)\Intel\oneAPI\

After this it is no longer necessary to have LLVM/bin on your Path.

If you have any issues contact me.