Actions speedup (#4)
* Try Speeding Up Builds

Think it compiles slower on 22.04 due to -O3 being the default release flag. Overriding to -O2 to see if it improves.

* Split Kepler

Seems to hangup on ASSET_Astro.cpp, splitting it and see if i see any improvements

* GCC is a lost cause on VM, need to find the offending clang flag

* Try 15

* Figured it out

Slow builds are due to openmp related bug in clang-14. Linux users should avoid clang-14 if compile time is issue. Changing actions to clang-13.

https://github.com/llvm/llvm-project/issues/56241